Moved field calculations to QuakeLibElement.cpp, added getElementIDs to ModelWorld, semi-finalized data structure for SlipMap class. Still untested
